fix(file): compare POSIX read counts portably
What changed: Normalize the probe read count to Long before comparing it with EOF.

Why: Android Native exposes the POSIX read count as Int, so the prior Int-to-Long comparison did not compile.

Compatibility and verification: EOF semantics unchanged. :file:compileKotlinAndroidNativeArm32 passes.

+++ file/src/posixMain/kotlin/maryk/file/File.kt
@@ -128,7 +128,7 @@ actual object File {
                         val count = read(fd, probePinned.addressOf(0), 1.convert())
                         if (count < 0 && errno == EINTR) continue
                         if (count < 0) return null
-                        return if (count == 0L) buffer else null
+                        return if (count.toLong() == 0L) buffer else null
                     }
                 }
                 return null
